+++ src/main/java/service/CreditCardService.java
@@ -5,6 +5,7 @@ package service;
 import domain.CreditCard;
 
 import java.util.List;
+import java.util.Map;
 
 public interface CreditCardService {
     List<CreditCard> getAllCreditsCardForUser(int userId);
@@ -15,5 +16,6 @@ public interface CreditCardService {
 
     String getTotalBalance(int userId);
 
-//    boolean userAccessToCreditCard(long creditCardId, int userId);
+    boolean userAccessToCreditCard(long creditCardId, int userId);
+    void putData(Map<String,String[]> dataWithForm, long creditCardId);
 }
