+++ src/main/java/service/PaymentService.java
@@ -1,7 +1,6 @@
 package service;
 
 import domain.Payment;
-
 import java.util.List;
 import java.util.ResourceBundle;
 
@@ -9,4 +8,5 @@ public interface PaymentService {
     List<Payment> getAllPaymentsByUserId(int userId);
     Payment getPaymentById(int paymentId);
     List<String> doPayment(long creditCardId, int paymentId, ResourceBundle lang);
+
 }
