+++ src/main/webapp/WEB-INF/css/style.css
@@ -1,0 +1,178 @@
+body {
+    margin:0;
+    font-family: 'Open Sans', sans-serif;
+    font-size: 14px;
+}
+
+*, *::before, *::after{
+    box-sizing: border-box;
+}
+
+.container{
+    width: 100%;
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+.header {
+    background-color: #2B2B2B;
+}
+
+.payment_system{
+    font-style: italic;
+    color: white;
+    font-size: large;
+    text-decoration: none;
+    opacity: .9;
+}
+
+.payment_system:hover{
+    text-decoration: underline;
+    opacity: 1;
+}
+
+.header__inner {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    border-bottom: 1px solid #515369;
+    padding: 25px 0;
+}
+
+.nav{
+    display: flex;
+    font-size: 13px;
+    font-weight: 700;
+    text-transform: uppercase;
+}
+
+.nav_link{
+    margin-left: 30px;
+    color: white;
+    opacity: .75;
+    transition: opacity .2s linear;
+    text-decoration: none;
+}
+
+.nav_link:first-child{
+    margin-left: 0;
+}
+
+.nav_link:hover{
+    text-decoration: underline;
+    opacity: 1;
+}
+
+.lang-switch a {
+    color:white;
+    text-decoration: none;
+}
+
+.lang-switch a:hover {
+    color:white;
+    text-decoration: underline;
+}
+
+.form_wrapper{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+
+.form_button:hover{
+    background-color: orange;
+}
+
+
+.in_form {
+    margin-top: 100px;
+}
+
+.in_form form{
+    background-color: grey;
+    padding: 20px;
+}
+
+.form_input {
+    margin-bottom: 5px;
+
+}
+
+.account_reference a{
+    text-decoration: none;
+    margin: 15px 5px;
+}
+
+.account_reference a:hover{
+    text-decoration: underline;
+}
+
+.greeting{
+    margin-top: 40px;
+    text-align: center;
+}
+
+.admin_main_page a{
+    padding-top: 20px;
+    margin: 15px 5px;
+}
+
+.pagination{
+    display:flex;
+    flex-direction: row;
+    justify-content: center;
+    text-decoration: none;
+}
+
+.pagination ul li{
+    list-style-type: none;
+    font-size: 30px;
+    padding: 5px;
+
+}
+
+.pagination ul li a{
+    text-decoration: none;
+    color: black;
+}
+
+.pagination ul li a:hover{
+    text-decoration: underline;
+    color: orange;
+}
+
+li .active {
+    color: rgb(245, 124, 3);
+}
+
+.footer {
+    display: flex;
+    justify-content: center;
+    background-color: #2B2B2B;
+    position: fixed;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    height: 80px;
+}
+
+.push {
+margin-bottom: 80px;
+}
+
+.footer_message {
+    color: white;
+    z-index: 10000;
+    font-size:medium;
+    font-style: italic;
+    text-align: center;
+    padding: 15px 10px;
+}
+
+
+.errors {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
