+++ src/main/webapp/WEB-INF/view/404.jsp
@@ -1,1 +1,2 @@
-NOT FOUND
+<jsp:include page="header.jsp"/>
+<h2>404</h2>
