+++ src/main/webapp/WEB-INF/view/account.jsp
@@ -4,18 +4,19 @@
     <head>
         <meta charset="UTF-8" />
         <title><c:out value="${lang.accountTitle}"/></title>
-    </head>
-    <body>
+
 
     <jsp:include page="header.jsp"/>
+    <div class="container">
 
-        <p>It's account</p>
+        <div class="account_reference">        
         <a href="/account/credit-cards"><c:out value="${lang.accountShowCreditCards}"/></a>
         <br>
 
         <a href="/account/payments"><c:out value="${lang.accountShowPayments}"/></a>
         <br>
-
+        </div>
+    </div>
 
     <jsp:include page="footer.jsp"/>
     </body>
