+++ src/main/webapp/WEB-INF/view/admin.jsp
@@ -4,14 +4,15 @@
     <head>
         <meta charset="UTF-8" />
         <title><c:out value="${lang.adminTitle}" /></title>
-    </head>
-    <body>
+
 
     <jsp:include page="header.jsp"/>
-
+    <div class="container">
+        <div class="admin_main_page">
        <a href="/admin/users"><c:out value="${lang.adminShowAllUsers}" /></a>
        <br>
-
+    </div>
+        </div>
 <jsp:include page="footer.jsp"/>
 
     </body>
