+++ src/main/webapp/WEB-INF/view/creditCards.jsp
@@ -4,11 +4,11 @@
     <head>
         <meta charset="UTF-8" />
         <title><c:out value="${lang.creditCardsTitle}"/></title>
-    </head>
-    <body>
+
 
     <jsp:include page="header.jsp"/>
 
+    <div class="container">
         <p><c:out value="${lang.creditCardsTotalBalance}"/> : ${totalBalance}</p>
        <c:forEach var="creditCard" items="${creditCards}">
                   <p><c:out value="${lang.creditCardsCardNumber}"/> : ${creditCard.idForUserView}</p>
@@ -17,7 +17,7 @@
                   <a href="/account/credit-cards/${creditCard.id}"><c:out value="${lang.creditCardsCardShowDetailsReference}"/></a>
                   <hr>
         </c:forEach>
-
+    </div>
 
 <jsp:include page="footer.jsp"/>
     </body>
