+++ src/main/webapp/WEB-INF/view/footer.jsp
@@ -1,5 +1,7 @@
-<footer>
+<div class="push"></div>
+<footer class="footer">
 
-<p><c:out value="${lang.footer}"/></p>
+    <p class="footer_message">Payment System final project</p>
+   
 
 </footer>
