+++ src/main/webapp/WEB-INF/view/payment.jsp
@@ -4,18 +4,17 @@
     <head>
         <meta charset="UTF-8" />
         <title><c:out value="${lang.paymentTitle}" /></title>
-    </head>
-    <body>
+
 
     <jsp:include page="header.jsp"/>
-
+    <div class="container">
        <c:forEach var="payment" items="${payments}">
                   <p><c:out value="${lang.paymentSum}" /> :${payment.sum}</p>
                   <p><c:out value="${lang.paymentDate}" /> :${payment.invoicedPaymentDate}</p>
                   <a href="/account/payments/${payment.id}"><c:out value="${lang.paymentShowDetails}" /></a>
                   <hr>
         </c:forEach>
-
+    <div class="container">   
 
 <jsp:include page="footer.jsp"/>
     </body>
