+++ src/main/webapp/WEB-INF/view/register.jsp
@@ -4,31 +4,41 @@
     <head>
         <meta charset="UTF-8" />
         <title><c:out value="${lang.registerTitle}"/></title>
-    </head>
-    <body>
+
 
     <jsp:include page="header.jsp"/>
 
-
-        ${serverError}
-        <c:forEach var="error" items="${errors}">
-         <p>${error}</p>
-        </c:forEach>
+    <div class="container">
+        <div class="form_wrapper">
+            <div class="in_form">
+        
+        
 
         <form method="POST">
-            <input type="text" placeholder="<c:out value="${lang.registerPlaceholderForName}"/>" name="name">
+            <input class="form_input" type="text" placeholder="<c:out value="${lang.registerPlaceholderForName}"/>" name="name">
             <br>
-            <input type="text" placeholder="<c:out value="${lang.registerPlaceholderForSurname}"/>" name="surname">
+            <input class="form_input" type="text" placeholder="<c:out value="${lang.registerPlaceholderForSurname}"/>" name="surname">
             <br>
-            <input type="password" placeholder="<c:out value="${lang.registerPlaceholderForPassword}"/>" name="password1">
+            <input class="form_input" type="password" placeholder="<c:out value="${lang.registerPlaceholderForPassword}"/>" name="password1">
             <br>
-            <input type="password" placeholder="<c:out value="${lang.registerPlaceholderForPassword}"/>" name="password2">
+            <input class="form_input" type="password" placeholder="<c:out value="${lang.registerPlaceholderForPassword}"/>" name="password2">
             <br>
-            <input type="email" placeholder="<c:out value="${lang.registerPlaceholderForEmail}"/>" name="email">
+            <input class="form_input" type="email" placeholder="<c:out value="${lang.registerPlaceholderForEmail}"/>" name="email">
             <br>
-            <input type="submit" value="<c:out value="${lang.registerButtonForRegister}"/>">
+            <input class="form_button" type="submit" value="<c:out value="${lang.registerButtonForRegister}"/>">
         </form>
 
+        
+            </div>
+            
+        </div>
+        <div class="errors">
+            ${serverError}
+            <c:forEach var="error" items="${errors}">
+             <p>${error}</p>
+            </c:forEach>
+            </div>
+    </div>
 
 
         <jsp:include page="footer.jsp"/>
